+++ crates/alpaca-trade-mock/src/handlers/mod.rs
@@ -1,5 +1,3 @@
-mod account;
 mod health;
 
-pub use account::get_account;
 pub use health::health;