+++ crates/alpaca-trade-mock/src/lib.rs
@@ -1,8 +1,6 @@
-pub mod admin;
 pub mod app;
 pub mod handlers;
 pub mod routes;
-pub mod state;
 
 use tokio::{net::TcpListener, task::JoinHandle};